+++ .npmrc
@@ -1,0 +1,2 @@
+save=true
+save-exact=true
